sidecar/requirements.txt
# Voice AI Sidecar - Dependencias
# Fase 1: MVP STT Local

# Framework Web
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# STT - Faster-Whisper (CTranslate2)
faster-whisper==1.0.3

# TTS - Piper (ONNX)
piper-tts>=1.2.0

# Audio Processing
pydub==0.25.1
soundfile==0.12.1
numpy>=1.24.0,<2.0.0

# HTTP Client (Gemini REST, chamadas externas)
httpx>=0.27.0

# Utilities
python-dotenv==1.0.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
